+++ src/ecdh.ts
@@ -1,5 +1,5 @@
-import { assert, isValidPrivateKey } from './validators';
-import { ecdhDerive } from './secp256k1';
+import { assert, isValidPrivateKey } from './helpers/validators';
+import { ecdhDerive } from './lib/secp256k1';
 
 export async function derive(
   privateKeyA: Buffer,
