+++ src/ecdsa.ts
@@ -1,11 +1,11 @@
-import { KeyPair } from './types';
-import { assert, isValidPrivateKey } from './validators';
+import { KeyPair } from './helpers/types';
+import { assert, isValidPrivateKey } from './helpers/validators';
 import {
   createPrivateKey,
   createPublicKey,
   ecdsaSign,
   ecdsaVerify,
-} from './secp256k1';
+} from './lib/secp256k1';
 
 export function generatePrivate() {
   return createPrivateKey();
