+++ src/ecies.ts
@@ -4,15 +4,20 @@ import { getPublic } from './ecdsa';
 import { randomBytes } from './random';
 import { hashSharedKey } from './sha2';
 import { getHmac } from './hmac';
-import { Encrypted, PreEncryptOpts } from './types';
-import { assert, isValidPrivateKey } from './validators';
-import { ENCRYPT_OP, SIGN_OP, VERIFY_OP, DECRYPT_OP } from './constants';
+import { Encrypted, PreEncryptOpts } from './helpers/types';
+import { assert, isValidPrivateKey } from './helpers/validators';
+import {
+  ENCRYPT_OP,
+  SIGN_OP,
+  VERIFY_OP,
+  DECRYPT_OP,
+} from './helpers/constants';
 
 async function getEncryptionKeys(privateKey: Buffer, publicKey: Buffer) {
-  const sharedKey: Buffer = await derive(privateKey, publicKey);
-  const hash: Uint8Array = await hashSharedKey(sharedKey);
-  const encryptionKey: Buffer = Buffer.from(hash.slice(0, 32));
-  const macKey: Buffer = Buffer.from(hash.slice(32));
+  const sharedKey = await derive(privateKey, publicKey);
+  const hash = await hashSharedKey(sharedKey);
+  const encryptionKey = Buffer.from(hash.slice(0, 32));
+  const macKey = Buffer.from(hash.slice(32));
   return { encryptionKey, macKey };
 }
 
@@ -22,7 +27,7 @@ async function handleEphemKeyPair(opts?: PreEncryptOpts) {
   while (!isValidPrivateKey(ephemPrivateKey)) {
     ephemPrivateKey = opts.ephemPrivateKey || randomBytes(32);
   }
-  const ephemPublicKey: Buffer = getPublic(ephemPrivateKey);
+  const ephemPublicKey = getPublic(ephemPrivateKey);
   return { ephemPrivateKey, ephemPublicKey };
 }
 
@@ -36,7 +41,7 @@ export async function encrypt(
     ephemPrivateKey,
     publicKeyTo
   );
-  const iv: Buffer = opts?.iv || randomBytes(16);
+  const iv = opts?.iv || randomBytes(16);
   const aesCbcEncrypt = getAes(ENCRYPT_OP);
   const ciphertext = await aesCbcEncrypt(iv, encryptionKey, msg);
   const dataToMac = Buffer.concat([iv, ephemPublicKey, ciphertext]);
@@ -54,7 +59,6 @@ export async function decrypt(
     privateKey,
     ephemPublicKey
   );
-
   const dataToMac = Buffer.concat([iv, ephemPublicKey, ciphertext]);
   const hmacSha256Verify = getHmac(VERIFY_OP);
   const macTest = await hmacSha256Verify(macKey, dataToMac, mac);
