+++ src/lib/secp256k1/typings.ts
@@ -1,23 +1,3 @@
-export interface Encrypted {
-  ciphertext: Buffer;
-  ephemPublicKey: Buffer;
-  iv: Buffer;
-  mac: Buffer;
-}
-
-export interface PreEncryptOpts extends Encrypted {
-  ephemPrivateKey: Buffer;
-}
-
-export interface KeyPair {
-  privateKey: Buffer;
-  publicKey: Buffer;
-}
-
-import * as eccryptoJS from './index';
-
-export type IEccrypto = typeof eccryptoJS;
-
 export type OutputModifier = Uint8Array | ((len: number) => Uint8Array);
 
 export interface ISecp256k1 {
